Changed handlers structure
