+++ src/debugger/cdp-proxy/CDPMessageHandlers/abstraction/SafariCDPMessageHandlerBase.ts
