+++ src/debugger/cdp-proxy/CDPMessageHandlers/implementation/chromeCordovaCDPMessageHandler.ts
@@ -1,16 +1,16 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for details.
 
-import * as url from "url";
-import { ProcessedCDPMessage, DispatchDirection } from "./CDPMessageHandlerBase";
-import { ChromeCDPMessageHandlerBase } from "./ChromeCDPMessageHandlerBase";
+import { ProcessedCDPMessage, DispatchDirection } from "../abstraction/CDPMessageHandlerBase";
+import { ChromeCDPMessageHandlerBase } from "../abstraction/ChromeCDPMessageHandlerBase";
 import { SourcemapPathTransformer } from "../../sourcemapPathTransformer";
 import { IProjectType } from "../../../../utils/cordovaProjectHelper";
 import { ICordovaAttachRequestArgs } from "../../../requestArgs";
 import { CDP_API_NAMES } from "../CDPAPINames";
-import { PlatformType } from "../../../cordovaDebugSession";
 
-export class ChromeIonicCDPMessageHandler extends ChromeCDPMessageHandlerBase {
+export class ChromeCordovaCDPMessageHandler extends ChromeCDPMessageHandlerBase {
+    private isSimulate: boolean;
+
     constructor(
         sourcemapPathTransformer: SourcemapPathTransformer,
         projectType: IProjectType,
@@ -18,17 +18,17 @@ export class ChromeIonicCDPMessageHandler extends ChromeCDPMessageHandlerBase {
     ) {
         super(sourcemapPathTransformer, projectType, args);
 
-        if (args.platform === PlatformType.Serve || args.ionicLiveReload) {
-            this.applicationPortPart = args.devServerPort ? `:${args.devServerPort}` : "";
-        }
         if (args.simulatePort) {
             this.applicationPortPart = `:${args.simulatePort}`;
+            this.isSimulate = true;
+        } else {
+            this.isSimulate = false;
         }
     }
 
     public processDebuggerCDPMessage(event: any): ProcessedCDPMessage {
         let dispatchDirection = DispatchDirection.FORWARD;
-        if (event.method === CDP_API_NAMES.DEBUGGER_SET_BREAKPOINT_BY_URL) {
+        if (event.method === CDP_API_NAMES.DEBUGGER_SET_BREAKPOINT_BY_URL && this.isSimulate) {
             event.params = this.fixSourcemapRegexp(event.params);
         }
 
@@ -45,7 +45,6 @@ export class ChromeIonicCDPMessageHandler extends ChromeCDPMessageHandlerBase {
             && event.params.url
             && event.params.url.startsWith(`http://${this.applicationServerAddress}`)
         ) {
-            this.tryToGetIonicDevServerPortFromURL(event.params.url);
             event.params = this.fixSourcemapLocation(event.params);
         }
 
@@ -65,22 +64,7 @@ export class ChromeIonicCDPMessageHandler extends ChromeCDPMessageHandlerBase {
             } else {
                 reqParams.url = "file://" + absoluteSourcePath;
             }
-        } else if (!(this.platform === PlatformType.Serve || (this.ionicLiveReload && this.debugRequestType === "launch"))) {
-            reqParams.url = "";
         }
         return reqParams;
-    }
-
-    private tryToGetIonicDevServerPortFromURL(sourceURL: string) {
-        if (this.ionicLiveReload && !this.applicationPortPart) {
-            try {
-                const devServerPort = url.parse(sourceURL).port;
-                if (devServerPort) {
-                    this.applicationPortPart = `:${devServerPort}`;
-                }
-            } catch (err) {
-                // do nothing, try to check another URL
-            }
-        }
     }
 }
